docapi/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0
django-filebrowser-no-grappelli==3.8.0
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
gunicorn==20.0.4
jsmin==2.2.2
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
